requirements.txt
Flask==2.0.2
pymongo==3.13.0
python-dotenv==0.15.0
python-snappy==0.6.0
zstandard==0.15.2